Completed 3rd Problem
Largest Prime Factor

+++ EulerTools/Helper.cs
@@ -4,7 +4,7 @@ namespace EulerTools
 {
     public static class EulerHelper
     {
-        public static void DisplayAnswer(int number)
+        public static void DisplayAnswer(long number)
         {
             Console.WriteLine(number);
             Console.ReadKey();
